Fix ~100 errors

+++ src/types/response_parameters.rs
@@ -1,5 +1,7 @@
+use serde::Deserialize;
+
 #[derive(Debug, Deserialize)]
 pub struct ResponseParameters{
-    pub migrate_to_chat_id: Option<i64>,
+    pub migrate_to_chat_id: Option<i32>,
     pub retry_after: Option<i32>
 }
